requirements.txt
python-dotenv==1.0.0
openai==0.27.7
langchain==0.0.181
Flask==2.3.2
shortuuid==1.0.11
fastapi==0.98.0
uvicorn==0.22.0